pyproject.toml
[project]
name = "communitymatcher"
version = "0.1.0"
description = "Community Finder / Recommendation Orchestrator using AWS Strands"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "psycopg2-binary>=2.9.0",
    "playwright>=1.40.0",
    "python-dateutil>=2.9.0",
    "tenacity>=8.2.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["community_matcher", "community_collector"]

[tool.pytest.ini_options]
testpaths = ["community_matcher/tests", "tests/collector"]
python_files = ["test_*.py"]
